+++ upcloud/testdata/upcloud_kubernetes/kubernetes_labels_s3.tf
@@ -29,4 +29,5 @@ resource "upcloud_kubernetes_cluster" "main" {
   name                    = "${var.basename}cluster"
   network                 = upcloud_network.main.id
   zone                    = var.zone
+  storage_encryption      = "data-at-rest"
 }